requirements.txt
pandas
inewave==0.0.65
python-dotenv
click
